react page and routes pages are created and del st
