+++ app/notes/(docs)/react/setup/page.tsx
