--- app/notes/docs/next/Links.ts
@@ -1,10 +1,0 @@
-export const nextLinks = [
-	{
-		links: "http://localhost:3000/notes/docs/next/setup",
-		link_title: "Setting up next",
-	},
-	{
-		links: "http://localhost:3000/notes/docs/next/intro",
-		link_title: "Introduction of next",
-	},
-];
