--- app/notes/docs/react/Links.ts
@@ -1,10 +1,0 @@
-export const reactLinks = [
-	{
-		links: "http://localhost:3000/notes/docs/react/setup",
-		link_title: "Setting up react",
-	},
-	{
-		links: "http://localhost:3000/notes/docs/react/intro",
-		link_title: "Introduction of React",
-	},
-];
